pyproject.toml
[project]

name = "pyri-example-browser"
version = "0.1.0"
description = "PyRI Teach Pendant WebUI Browser Example"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
keywords = ["pyodide", "pyri-project", "pyri-webui", "pyri-webui-browser-package", "robotics"]
authors = [
    {name = "John Wason", email = "wason@wasontech.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering"
]

dependencies = [
    'pyri-common',        
    'importlib-resources==5.13.0'
]

[project.urls]
homepage="http://pyri.tech"
repository="https://github.com/pyri-project/pyri-example-browser"

[project.entry-points."pyri.plugins.webui_browser_panel"]
pyri-robotics-browser   = "pyri_example_browser.panels.example_panels:get_webui_browser_panel_factory"

[project.entry-points."pyri.plugins.webui_browser_component"]
pyri-robotics-browser = "pyri_example_browser.components.example_components:get_webui_browser_component_factory"

[build-system]
build-backend = 'setuptools.build_meta'
requires = [
    'setuptools',
    'toml',
]

[tool.setuptools.package-data]
'pyri_example_browser.panels' = ['*.html']
'pyri_example_browser.components' = ['*.html']